Add EditProfile method to User class
Introduced a new `EditProfile` method in the `User` class to allow updating the `Name`, `Family`, and `Gender` properties. This method takes three parameters (`name`, `family`, and `gender`) and assigns their values to the corresponding properties. No changes were made to the `ChangePassword` method, which is included in the diff for context.