Remove UserAddressController and related functionality
The `UserAddressController.cs` file has been completely removed.
This includes the deletion of all `using` directives, the
`namespace Shop.Api.Controllers` declaration, and the
`UserAddressController` class itself. The class previously
provided an HTTP GET endpoint (`GetList`) to fetch a list of
user addresses using the `IUserAddressFacade`. This endpoint
and its associated logic have been removed as part of this
change.